feat(mgmt): split grpc server convert code
