+++ mgmt/src/grpc/mod.rs
@@ -4,5 +4,6 @@
 //! Dataplane gRPC handling module.
 //! Implements gRPC request reception and response building.
 
+pub mod converter;
 pub mod server;
 pub mod test;
