+++ mgmt/src/grpc/test.rs
@@ -3,14 +3,14 @@
 
 #[cfg(test)]
 mod tests {
-    use std::collections::HashMap;
-    use std::sync::Arc;
-    use tokio::sync::RwLock;
-    // Import proto-generated types
-    use crate::frr::frrmi::FrrMi;
-    use crate::grpc::server::BasicConfigManager;
-    use crate::models::external::configdb::gwconfigdb::GwConfigDatabase;
     use gateway_config::GatewayConfig;
+    use std::collections::HashMap;
+    use std::convert::TryFrom;
+
+    // Import converter module
+    use crate::grpc::converter;
+    use crate::models::internal::device::DeviceConfig;
+    use crate::models::internal::interfaces::interface::InterfaceConfig;
 
     // Helper function to create a test GatewayConfig
     fn create_test_gateway_config() -> GatewayConfig {
@@ -187,31 +187,19 @@ mod tests {
         }
     }
 
-    use crate::frr::frrmi::tests::fake_frr_agent;
+    use crate::frr::frrmi::open_unix_sock_async;
 
     #[tokio::test]
     async fn test_convert_to_grpc_config() {
         // Create a mock database
-        let config_db = Arc::new(RwLock::new(GwConfigDatabase::new()));
-
-        // create faked frr-agent. This should not be needed in this test.
-        let frr_agent = fake_frr_agent("/tmp/grpc/frr-agent.sock").await;
-
-        // create frrmi
-        let frrmi = FrrMi::new("/tmp/frrmi.sock", "/tmp/grpc/frr-agent.sock")
-            .await
-            .unwrap();
-
-        frr_agent.abort();
-
-        // Create the manager
-        let manager = BasicConfigManager::new(Arc::clone(&config_db), frrmi);
-
+        #[allow(unused_variables)]
+        let sock = open_unix_sock_async("/tmp/frr-agent.sock").expect("Should succeed");
         // Create test data
         let grpc_config = create_test_gateway_config();
 
         // Call the conversion function (gRPC -> ExternalConfig)
-        let result = manager.convert_from_grpc_config(&grpc_config).await;
+        // Using standalone function instead of manager method
+        let result = converter::convert_from_grpc_config(&grpc_config).await;
 
         // Verify result
         assert!(
@@ -222,7 +210,8 @@ mod tests {
         let external_config = result.unwrap();
 
         // Call the conversion function (ExternalConfig -> gRPC)
-        let result = manager.convert_to_grpc_config(&external_config).await;
+        // Using standalone function instead of manager method
+        let result = converter::convert_to_grpc_config(&external_config).await;
 
         // Verify result
         assert!(
@@ -259,6 +248,29 @@ mod tests {
         assert_eq!(
             converted_device.loglevel, original_device.loglevel,
             "Device log level mismatch"
+        );
+
+        // --- Test TryFrom implementation ---
+        // Convert device using TryFrom
+        let device_result = DeviceConfig::try_from(original_device);
+        assert!(
+            device_result.is_ok(),
+            "TryFrom conversion for device failed"
+        );
+        let internal_device = device_result.unwrap();
+
+        // Convert back to gRPC using TryFrom
+        let device_grpc_result = gateway_config::Device::try_from(&internal_device);
+        assert!(
+            device_grpc_result.is_ok(),
+            "TryFrom conversion back to gRPC device failed"
+        );
+        let device_grpc = device_grpc_result.unwrap();
+
+        // Verify conversion via TryFrom
+        assert_eq!(
+            device_grpc.hostname, original_device.hostname,
+            "Device hostname mismatch using TryFrom"
         );
 
         // --- UNDERLAY CONFIGURATION TESTS ---
@@ -347,6 +359,24 @@ mod tests {
                     "VLAN ID mismatch for interface {name}",
                 );
             }
+
+            // Test TryFrom for interface
+            let interface_result = InterfaceConfig::try_from(*original_iface);
+            assert!(
+                interface_result.is_ok(),
+                "TryFrom conversion for interface failed for {}",
+                name
+            );
+
+            let internal_interface = interface_result.unwrap();
+
+            // Test converting back
+            let interface_grpc_result = gateway_config::Interface::try_from(&internal_interface);
+            assert!(
+                interface_grpc_result.is_ok(),
+                "TryFrom conversion back to gRPC interface failed for {}",
+                name
+            );
         }
 
         // --- BGP CONFIGURATION TESTS ---
@@ -563,5 +593,74 @@ mod tests {
                 }
             }
         }
+    }
+
+    #[tokio::test]
+    async fn test_tryfrom_conversions() {
+        // Create test data with specific components
+        let device = gateway_config::Device {
+            driver: 0,   // Kernel
+            loglevel: 2, // INFO
+            hostname: "test-device".to_string(),
+            ports: Vec::new(),
+            eal: None,
+        };
+
+        let interface = gateway_config::Interface {
+            name: "eth0".to_string(),
+            ipaddr: "192.168.1.1/24".to_string(),
+            r#type: 0, // Ethernet
+            role: 0,   // Fabric
+            vlan: None,
+            macaddr: Some("00:11:22:33:44:55".to_string()),
+            system_name: Some("".to_string()),
+            ospf: None,
+        };
+
+        // Test DeviceConfig TryFrom
+        let device_config_result = DeviceConfig::try_from(&device);
+        assert!(
+            device_config_result.is_ok(),
+            "TryFrom for DeviceConfig failed"
+        );
+        let device_config = device_config_result.unwrap();
+
+        // Verify conversion result
+        assert_eq!(device_config.settings.hostname, "test-device");
+
+        // Convert back using TryFrom
+        let device_back_result = gateway_config::Device::try_from(&device_config);
+        assert!(device_back_result.is_ok(), "TryFrom back to Device failed");
+        let device_back = device_back_result.unwrap();
+
+        // Verify round trip conversion
+        assert_eq!(device_back.hostname, device.hostname);
+        assert_eq!(device_back.driver, device.driver);
+        assert_eq!(device_back.loglevel, device.loglevel);
+
+        // Test InterfaceConfig TryFrom
+        let interface_config_result = InterfaceConfig::try_from(&interface);
+        assert!(
+            interface_config_result.is_ok(),
+            "TryFrom for InterfaceConfig failed"
+        );
+        let interface_config = interface_config_result.unwrap();
+
+        // Verify conversion result
+        assert_eq!(interface_config.name, "eth0");
+        assert!(!interface_config.addresses.is_empty());
+
+        // Convert back using TryFrom
+        let interface_back_result = gateway_config::Interface::try_from(&interface_config);
+        assert!(
+            interface_back_result.is_ok(),
+            "TryFrom back to Interface failed"
+        );
+        let interface_back = interface_back_result.unwrap();
+
+        // Verify round trip conversion
+        assert_eq!(interface_back.name, interface.name);
+        assert_eq!(interface_back.r#type, interface.r#type);
+        assert!(!interface_back.ipaddr.is_empty());
     }
 }
